Favoriting checkpoint completed

+++ app/controllers/favorites_controller.rb
@@ -4,6 +4,8 @@ class FavoritesController < ApplicationController
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: post)
 
+    authorize favorite
+
     if favorite.save
       flash[:notice] = "Post has been favorited."
       redirect_to [post.topic, post]
@@ -15,7 +17,9 @@ class FavoritesController < ApplicationController
 
   def destroy
     post = Post.find(params[:post_id])
-    favorite = post.favorites.where(user: current_user, post: post)
+    favorite = current_user.favorites.find(params[:id])
+
+    authorize favorite
 
     if favorite.destroy
       flash[:notice] = "Post has been successfully unfavorited."
